src: move ipyvizzulib into chartlib
