+++ src/ipyvizzu/chartlib/templates/ipyvizzu.js
