+++ src/ipyvizzu/jupyter/chart.py
@@ -1,0 +1,7 @@
+"""A module for working with Vizzu charts."""
+
+from ipyvizzu.chartlib.chart import ChartLib
+
+
+class Chart(ChartLib):
+    pass
